+++ bin/release.js
@@ -22,6 +22,10 @@ const config = {
     gitEmail: 'ds@gitmax.tech',
 };
 
+function escapeShellChars(str) {
+    return str.replace(/["`$]/g, '\\$&');
+}
+
 function setupGit() {
     shell.exec(`git config user.name "${config.gitUsername}"`, execOptions);
     shell.exec(`git config user.email "${config.gitEmail}"`, execOptions);
@@ -159,7 +163,7 @@ async function releaseRoot() {
     await git.add(config.changelogPath, cwd);
 
     logger.log('=> Commit changed files');
-    shell.exec('git commit -n -m "chore: publish root package"', execOptions);
+    shell.exec('git commit -n -m "chore: publish root package [skip ci]"', execOptions);
 
     // копирую package.json в сборку корневого пакета
     shell.exec('cp package.json dist/package.json', execOptions);
@@ -192,7 +196,9 @@ async function releaseRoot() {
 
         logger.log('=> Create github release');
         shell.exec(
-            `gh release create ${nextReleaseTag} --title "${nextReleaseTag}" --notes "${notes}" --target master`,
+            `gh release create ${nextReleaseTag} --title "${nextReleaseTag}" --notes "${escapeShellChars(
+                notes,
+            )}" --target master`,
             execOptions,
         );
 
@@ -218,7 +224,7 @@ async function releasePackages() {
     shell.exec('git add .', execOptions);
     // Не добавляем .npmrc в коммит.
     shell.exec('git reset .npmrc', execOptions);
-    shell.exec('git commit -n -m "chore: publish packages"', execOptions);
+    shell.exec('git commit -n -m "chore: publish packages [skip ci]"', execOptions);
 
     logger.log('=> Push changes');
     shell.exec(