+++ packages/attach/src/component.screenshots.test.tsx
@@ -63,6 +63,8 @@ describe('Attach | interactions tests', () => {
         } catch (error) {
             // eslint-disable-next-line no-console
             console.error(error.message);
+
+            throw error;
         } finally {
             await closeBrowser({ browser, context, page });
         }